requirements.txt
python-dotenv
pandas
telethon